#Refactoring

+++ src/Capitan360.Application/Services/CompanyServices/CompanyCommissions/Dtos/CompanyCommissionsDto.cs
@@ -5,10 +5,10 @@ public class CompanyCommissionsDto
     public int Id { get; set; }
     public int CompanyId { get; set; }
     public string? CompanyName { get; set; }
-    public decimal CommissionFromCaptainCargoWebSite { get; set; }
-    public decimal CommissionFromCompanyWebSite { get; set; }
-    public decimal CommissionFromCaptainCargoWebService { get; set; }
-    public decimal CommissionFromCompanyWebService { get; set; }
-    public decimal CommissionFromCaptainCargoPanel { get; set; }
-    public decimal CommissionFromCaptainCargoDesktop { get; set; }
+    public long CommissionFromCaptainCargoWebSite { get; set; }
+    public long CommissionFromCompanyWebSite { get; set; }
+    public long CommissionFromCaptainCargoWebService { get; set; }
+    public long CommissionFromCompanyWebService { get; set; }
+    public long CommissionFromCaptainCargoPanel { get; set; }
+    public long CommissionFromCaptainCargoDesktop { get; set; }
 }
